requirements.txt
requests
SQLAlchemy
pandas
docker
python-dotenv
logging
datetime